feat(settings): migrate fields into 5-section IA (PR2)
Decompose TabbedConfigForm into settingsContent.tsx (field registry,
helpers, section renderers — Scrape/Network/Download/PersonSync split to
fit the new IA) and SettingsForm.tsx (5 SectionAnchor x Subsection
layout: Data Sources / Rate Limiting / Extraction Rules / Paths /
System). Route is now thin, no tab chrome, no settingsActiveTab slice.
