+++ src/renderer/src/components/settings/settingsContent.tsx
@@ -1,0 +1,816 @@
+import { isSharedDirectoryMode } from "@shared/assetNaming";
+import type { Configuration } from "@shared/config";
+import { TRANSLATION_TARGET_OPTIONS } from "@shared/enums";
+import { DEFAULT_LLM_BASE_URL } from "@shared/llm";
+import type { NamingPreviewItem } from "@shared/types";
+import { useQuery } from "@tanstack/react-query";
+import { Loader2, RotateCcw } from "lucide-react";
+import { useEffect, useMemo, useRef, useState } from "react";
+import type { FieldValues } from "react-hook-form";
+import { useFormContext, useWatch } from "react-hook-form";
+import { toast } from "sonner";
+import { ipc } from "@/client/ipc";
+import {
+  BaseField,
+  BoolField,
+  ChipArrayFieldWrapper,
+  CookieFieldWrapper,
+  DurationFieldWrapper,
+  EnumField,
+  type EnumOption,
+  NumberField,
+  OrderedSiteFieldWrapper,
+  PathFieldWrapper,
+  PromptFieldWrapper,
+  SecretField,
+  ShortcutField,
+  TextField,
+  UrlField,
+} from "@/components/config-form/FieldRenderer";
+import { Button } from "@/components/ui/Button";
+import { Form, FormControl } from "@/components/ui/Form";
+import { Switch } from "@/components/ui/Switch";
+
+export { Form };
+
+// ── Constants ──
+
+const PROXY_TYPE_OPTIONS = ["none", "http", "https", "socks5"];
+const TRANSLATE_ENGINE_OPTIONS: EnumOption[] = [
+  { value: "openai", label: "LLM 翻译" },
+  { value: "google", label: "Google 翻译（免费）" },
+];
+const LANGUAGE_OPTIONS = [...TRANSLATION_TARGET_OPTIONS];
+const ACTOR_OVERVIEW_SOURCE_OPTIONS = ["official", "avjoho", "avbase"];
+const ACTOR_IMAGE_SOURCE_OPTIONS = ["local", "gfriends", "official", "avbase"];
+const PART_STYLE_OPTIONS: EnumOption[] = [
+  { value: "RAW", label: "保持原始后缀" },
+  { value: "CD", label: "统一为 CD1 / CD2" },
+  { value: "PART", label: "统一为 PART1 / PART2" },
+  { value: "DISC", label: "统一为 DISC1 / DISC2" },
+];
+const ASSET_NAMING_OPTIONS: EnumOption[] = [
+  { value: "fixed", label: "固定命名" },
+  { value: "followVideo", label: "跟随影片文件名" },
+];
+const NFO_NAMING_OPTIONS: EnumOption[] = [
+  { value: "both", label: "同时生成两种" },
+  { value: "movie", label: "仅 movie.nfo" },
+  { value: "filename", label: "仅 文件名.nfo" },
+];
+
+export const NAMING_TEMPLATE_DESCRIPTION =
+  "可用占位符：{actor} {actorFallbackPrefix} {number} {date} {title} {originaltitle} {studio} {publisher}";
+
+// ── Field registry for search/filter ──
+
+export interface FieldEntry {
+  key: string;
+  /** Display label (Chinese). */
+  label: string;
+  /** Top-level anchor section the field belongs to in the new IA. */
+  anchor: "dataSources" | "rateLimiting" | "extractionRules" | "paths" | "system";
+}
+
+export const FIELD_REGISTRY: FieldEntry[] = [
+  // paths
+  { key: "paths.mediaPath", label: "媒体目录", anchor: "paths" },
+  { key: "paths.actorPhotoFolder", label: "演员头像库目录", anchor: "paths" },
+  { key: "paths.softlinkPath", label: "软链接目录", anchor: "paths" },
+  { key: "paths.successOutputFolder", label: "成功输出目录", anchor: "paths" },
+  { key: "paths.failedOutputFolder", label: "失败输出目录", anchor: "paths" },
+  { key: "paths.outputSummaryPath", label: "概览统计目录", anchor: "paths" },
+  { key: "paths.sceneImagesFolder", label: "剧照目录名", anchor: "paths" },
+  { key: "paths.configDirectory", label: "配置文件目录", anchor: "paths" },
+  // scrape — sites live in Data Sources
+  { key: "scrape.sites", label: "启用站点与优先级", anchor: "dataSources" },
+  // scrape — pacing lives in Rate Limiting
+  { key: "scrape.threadNumber", label: "并发线程数", anchor: "rateLimiting" },
+  { key: "scrape.javdbDelaySeconds", label: "JavDB 请求延迟(秒)", anchor: "rateLimiting" },
+  { key: "scrape.restAfterCount", label: "连续刮削后休息(条数)", anchor: "rateLimiting" },
+  { key: "scrape.restDuration", label: "休息时长", anchor: "rateLimiting" },
+  // network — connection in Rate Limiting
+  { key: "network.proxyType", label: "代理类型", anchor: "rateLimiting" },
+  { key: "network.proxy", label: "代理地址", anchor: "rateLimiting" },
+  { key: "network.useProxy", label: "启用代理", anchor: "rateLimiting" },
+  { key: "network.timeout", label: "超时时间(秒)", anchor: "rateLimiting" },
+  { key: "network.retryCount", label: "重试次数", anchor: "rateLimiting" },
+  // network — site cookies in Data Sources
+  { key: "network.javdbCookie", label: "JavDB 凭证", anchor: "dataSources" },
+  { key: "network.javbusCookie", label: "JavBus 凭证", anchor: "dataSources" },
+  // download
+  { key: "download.downloadThumb", label: "下载横版缩略图", anchor: "extractionRules" },
+  { key: "download.downloadPoster", label: "下载海报", anchor: "extractionRules" },
+  { key: "download.tagBadges", label: "封面标签角标", anchor: "extractionRules" },
+  { key: "download.downloadFanart", label: "下载背景图", anchor: "extractionRules" },
+  { key: "download.downloadSceneImages", label: "下载剧照", anchor: "extractionRules" },
+  { key: "download.downloadTrailer", label: "下载预告片", anchor: "extractionRules" },
+  { key: "download.generateNfo", label: "生成 NFO", anchor: "extractionRules" },
+  { key: "download.nfoNaming", label: "NFO 文件命名", anchor: "extractionRules" },
+  { key: "download.keepThumb", label: "保留已有横版缩略图", anchor: "extractionRules" },
+  { key: "download.keepPoster", label: "保留已有海报", anchor: "extractionRules" },
+  { key: "download.keepFanart", label: "保留已有背景图", anchor: "extractionRules" },
+  { key: "download.keepSceneImages", label: "保留已有剧照", anchor: "extractionRules" },
+  { key: "download.keepTrailer", label: "保留已有预告片", anchor: "extractionRules" },
+  { key: "download.keepNfo", label: "保留已有 NFO", anchor: "extractionRules" },
+  // naming
+  { key: "naming.folderTemplate", label: "文件夹模板", anchor: "extractionRules" },
+  { key: "naming.fileTemplate", label: "文件名模板", anchor: "extractionRules" },
+  { key: "naming.assetNamingMode", label: "附属文件命名", anchor: "extractionRules" },
+  { key: "naming.nfoTitleTemplate", label: "NFO 标题模板", anchor: "extractionRules" },
+  { key: "naming.actorNameMax", label: "演员名最大数量", anchor: "extractionRules" },
+  { key: "naming.actorNameMore", label: "演员名超出后缀", anchor: "extractionRules" },
+  { key: "naming.actorFallbackToStudio", label: "演员为空时使用片商或卖家", anchor: "extractionRules" },
+  { key: "naming.releaseRule", label: "发行日期格式", anchor: "extractionRules" },
+  { key: "naming.folderNameMax", label: "文件夹名最大长度", anchor: "extractionRules" },
+  { key: "naming.fileNameMax", label: "文件名最大长度", anchor: "extractionRules" },
+  { key: "naming.cnwordStyle", label: "中文字幕标记", anchor: "extractionRules" },
+  { key: "naming.umrStyle", label: "UMR 标记", anchor: "extractionRules" },
+  { key: "naming.leakStyle", label: "流出标记", anchor: "extractionRules" },
+  { key: "naming.uncensoredStyle", label: "无码标记", anchor: "extractionRules" },
+  { key: "naming.censoredStyle", label: "有码标记", anchor: "extractionRules" },
+  { key: "naming.partStyle", label: "分盘样式", anchor: "extractionRules" },
+  // translate
+  { key: "translate.enableTranslation", label: "启用内容翻译", anchor: "dataSources" },
+  { key: "translate.engine", label: "翻译引擎", anchor: "dataSources" },
+  { key: "translate.llmModelName", label: "LLM 模型名称", anchor: "dataSources" },
+  { key: "translate.llmApiKey", label: "LLM 密钥", anchor: "dataSources" },
+  { key: "translate.llmBaseUrl", label: "LLM API 地址", anchor: "dataSources" },
+  { key: "translate.llmPrompt", label: "LLM 翻译提示词", anchor: "dataSources" },
+  { key: "translate.llmTemperature", label: "LLM 温度", anchor: "dataSources" },
+  { key: "translate.llmMaxRetries", label: "LLM 最大重试次数", anchor: "dataSources" },
+  { key: "translate.llmMaxRequestsPerSecond", label: "LLM 每秒最大请求数", anchor: "dataSources" },
+  { key: "translate.targetLanguage", label: "目标语言", anchor: "dataSources" },
+  // personSync
+  { key: "personSync.personOverviewSources", label: "人物简介来源顺序", anchor: "dataSources" },
+  { key: "personSync.personImageSources", label: "人物头像来源顺序", anchor: "dataSources" },
+  // jellyfin
+  { key: "jellyfin.url", label: "Jellyfin 服务器地址", anchor: "dataSources" },
+  { key: "jellyfin.apiKey", label: "Jellyfin API Key", anchor: "dataSources" },
+  { key: "jellyfin.userId", label: "Jellyfin 用户 ID", anchor: "dataSources" },
+  { key: "jellyfin.refreshPersonAfterSync", label: "同步后刷新人物 (Jellyfin)", anchor: "dataSources" },
+  { key: "jellyfin.lockOverviewAfterSync", label: "同步后锁定人物简介 (Jellyfin)", anchor: "dataSources" },
+  // emby
+  { key: "emby.url", label: "Emby 服务器地址", anchor: "dataSources" },
+  { key: "emby.apiKey", label: "Emby API Key", anchor: "dataSources" },
+  { key: "emby.userId", label: "Emby 用户 ID", anchor: "dataSources" },
+  { key: "emby.refreshPersonAfterSync", label: "同步后刷新人物 (Emby)", anchor: "dataSources" },
+  // shortcuts
+  { key: "shortcuts.startOrStopScrape", label: "开始/停止刮削", anchor: "system" },
+  { key: "shortcuts.retryScrape", label: "重新刮削", anchor: "system" },
+  { key: "shortcuts.deleteFile", label: "删除文件", anchor: "system" },
+  { key: "shortcuts.deleteFileAndFolder", label: "删除文件及文件夹", anchor: "system" },
+  { key: "shortcuts.openFolder", label: "打开所在目录", anchor: "system" },
+  { key: "shortcuts.editNfo", label: "编辑 NFO", anchor: "system" },
+  { key: "shortcuts.playVideo", label: "播放视频", anchor: "system" },
+  // ui
+  { key: "ui.showLogsPanel", label: "显示日志面板", anchor: "system" },
+  { key: "ui.useCustomTitleBar", label: "使用自定义标题栏", anchor: "system" },
+  { key: "ui.hideDock", label: "隐藏 Dock 图标", anchor: "system" },
+  { key: "ui.hideMenu", label: "隐藏菜单栏", anchor: "system" },
+  { key: "ui.hideWindowButtons", label: "隐藏窗口按钮", anchor: "system" },
+  // behavior
+  { key: "behavior.successFileMove", label: "成功后移动文件", anchor: "system" },
+  { key: "behavior.failedFileMove", label: "失败后移动文件", anchor: "system" },
+  { key: "behavior.successFileRename", label: "成功后重命名文件", anchor: "system" },
+  { key: "behavior.deleteEmptyFolder", label: "删除空文件夹", anchor: "system" },
+  { key: "behavior.scrapeSoftlinkPath", label: "刮削软链接目录", anchor: "system" },
+  { key: "behavior.saveLog", label: "保存日志到文件", anchor: "system" },
+];
+
+export const SECTION_DESCRIPTIONS: Record<FieldEntry["anchor"], string> = {
+  dataSources: "刮削站点、翻译、人物同步服务的数据来源与凭证",
+  rateLimiting: "并发、延迟、重试、代理等节奏与连接控制",
+  extractionRules: "抓取策略、命名模板、资源下载与 NFO",
+  paths: "媒体库、头像库、输出与配置目录",
+  system: "界面、快捷键、刮削后的文件行为",
+};
+
+export const SECTION_LABELS: Record<FieldEntry["anchor"], string> = {
+  dataSources: "数据源",
+  rateLimiting: "速率与限流",
+  extractionRules: "提取规则",
+  paths: "目录与路径",
+  system: "系统",
+};
+
+// ── Config value helpers ──
+
+function getNestedValue(obj: Record<string, unknown>, path: string): unknown {
+  const parts = path.split(".");
+  let cursor: unknown = obj;
+  for (const part of parts) {
+    if (cursor == null || typeof cursor !== "object") return undefined;
+    cursor = (cursor as Record<string, unknown>)[part];
+  }
+  return cursor;
+}
+
+function setNestedValue(obj: Record<string, unknown>, path: string, value: unknown): void {
+  const parts = path.split(".");
+  let cursor: Record<string, unknown> = obj;
+  for (const part of parts.slice(0, -1)) {
+    if (!cursor[part] || typeof cursor[part] !== "object") cursor[part] = {};
+    cursor = cursor[part] as Record<string, unknown>;
+  }
+  const tail = parts.at(-1);
+  if (tail) cursor[tail] = value;
+}
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+export function flattenConfig(data: Record<string, unknown>): Record<string, unknown> {
+  const flat: Record<string, unknown> = {};
+  for (const entry of FIELD_REGISTRY) {
+    flat[entry.key] = getNestedValue(data, entry.key);
+  }
+
+  const siteConfigs = getNestedValue(data, "scrape.siteConfigs");
+  if (isRecord(siteConfigs)) {
+    for (const [site, config] of Object.entries(siteConfigs)) {
+      if (!isRecord(config)) {
+        continue;
+      }
+      if ("customUrl" in config) {
+        flat[`scrape.siteConfigs.${site}.customUrl`] = config.customUrl;
+      }
+    }
+  }
+
+  return flat;
+}
+
+export function unflattenConfig(flat: Record<string, unknown>): Record<string, unknown> {
+  const nested: Record<string, unknown> = {};
+  for (const [key, value] of Object.entries(flat)) {
+    if (value !== undefined) setNestedValue(nested, key, value);
+  }
+  return nested;
+}
+
+const NAMING_PREVIEW_FIELD_KEYS = [
+  "naming.folderTemplate",
+  "naming.fileTemplate",
+  "naming.assetNamingMode",
+  "naming.actorNameMax",
+  "naming.actorNameMore",
+  "naming.actorFallbackToStudio",
+  "naming.releaseRule",
+  "naming.folderNameMax",
+  "naming.fileNameMax",
+  "naming.cnwordStyle",
+  "naming.umrStyle",
+  "naming.leakStyle",
+  "naming.uncensoredStyle",
+  "naming.censoredStyle",
+  "naming.partStyle",
+  "download.nfoNaming",
+  "download.downloadSceneImages",
+  "behavior.successFileMove",
+  "behavior.successFileRename",
+] as const;
+
+export function buildNamingPreviewConfig(values: Record<string, unknown>): Partial<Configuration> {
+  const flat: Record<string, unknown> = {};
+  for (const key of NAMING_PREVIEW_FIELD_KEYS) {
+    const value = values[key] ?? getNestedValue(values, key);
+    if (value !== undefined) {
+      flat[key] = value;
+    }
+  }
+
+  return unflattenConfig(flat) as Partial<Configuration>;
+}
+
+function toStringArray(value: unknown): string[] {
+  if (!Array.isArray(value)) {
+    return [];
+  }
+  return value.filter((item): item is string => typeof item === "string");
+}
+
+function toSiteOptions(value: unknown): string[] {
+  if (!Array.isArray(value)) {
+    return [];
+  }
+
+  const outputs: string[] = [];
+  for (const item of value) {
+    if (typeof item === "string") {
+      outputs.push(item);
+      continue;
+    }
+    if (isRecord(item) && typeof item.site === "string") {
+      outputs.push(item.site);
+    }
+  }
+  return outputs;
+}
+
+type CrawlerSiteInfo = {
+  site: string;
+  name: string;
+  enabled: boolean;
+  native: boolean;
+};
+
+export function useCrawlerSiteOptions(flatDefaults: Record<string, unknown>): string[] {
+  const sitesQ = useQuery({
+    queryKey: ["crawler", "sites"],
+    queryFn: async () => {
+      const result = (await ipc.crawler.listSites()) as { sites: CrawlerSiteInfo[] };
+      return result.sites;
+    },
+    staleTime: 60_000,
+  });
+
+  return useMemo(() => {
+    const fromApi = toSiteOptions(sitesQ.data);
+    const fromConfig = toStringArray(flatDefaults["scrape.sites"]);
+    return Array.from(new Set([...fromApi, ...fromConfig]));
+  }, [sitesQ.data, flatDefaults]);
+}
+
+// ── Section renderers ──
+
+export function PathsSection() {
+  return (
+    <>
+      <PathFieldWrapper name="paths.mediaPath" label="媒体目录" isDirectory />
+      <PathFieldWrapper
+        name="paths.actorPhotoFolder"
+        label="演员头像库目录"
+        description="建议放在媒体库目录下。这里的头像会优先使用；如果希望优先使用在线头像，请在“人物头像来源顺序”中调整“本地”的位置或移除它。"
+        isDirectory
+      />
+      <PathFieldWrapper name="paths.softlinkPath" label="软链接目录" isDirectory />
+      <PathFieldWrapper name="paths.successOutputFolder" label="成功输出目录" isDirectory />
+      <PathFieldWrapper name="paths.failedOutputFolder" label="失败输出目录" isDirectory />
+      <PathFieldWrapper
+        name="paths.outputSummaryPath"
+        label="概览统计目录"
+        description="留空则使用成功输出目录"
+        isDirectory
+      />
+      <TextField name="paths.sceneImagesFolder" label="剧照目录名" />
+      <PathFieldWrapper name="paths.configDirectory" label="配置文件目录" isDirectory />
+    </>
+  );
+}
+
+interface ScrapeSitesSectionProps {
+  siteOptions: string[];
+}
+
+export function ScrapeSitesSection({ siteOptions }: ScrapeSitesSectionProps) {
+  return (
+    <OrderedSiteFieldWrapper
+      name="scrape.sites"
+      label="启用站点与优先级"
+      description="勾选启用站点，上下移动调整优先级。未勾选的站点不会参与刮削。"
+      options={siteOptions}
+    />
+  );
+}
+
+export function ScrapePacingSection() {
+  return (
+    <>
+      <NumberField name="scrape.threadNumber" label="并发线程数" min={1} max={128} />
+      <NumberField name="scrape.javdbDelaySeconds" label="JavDB 请求延迟(秒)" min={0} max={120} />
+      <NumberField name="scrape.restAfterCount" label="连续刮削后休息(条数)" min={1} max={500} />
+      <DurationFieldWrapper name="scrape.restDuration" label="休息时长" />
+    </>
+  );
+}
+
+export function NetworkConnectionSection() {
+  return (
+    <>
+      <EnumField name="network.proxyType" label="代理类型" options={PROXY_TYPE_OPTIONS} />
+      <TextField name="network.proxy" label="代理地址" />
+      <BoolField name="network.useProxy" label="启用代理" />
+      <NumberField name="network.timeout" label="超时时间(秒)" min={1} max={300} />
+      <NumberField name="network.retryCount" label="重试次数" min={0} max={10} />
+    </>
+  );
+}
+
+export function NetworkCookiesSection() {
+  return (
+    <>
+      <CookieFieldWrapper name="network.javdbCookie" label="JavDB 凭证" />
+      <CookieFieldWrapper name="network.javbusCookie" label="JavBus 凭证" />
+    </>
+  );
+}
+
+export function AssetDownloadsSection() {
+  const form = useFormContext<FieldValues>();
+  const [downloadThumb, downloadPoster, downloadFanart, downloadSceneImages, downloadTrailer] = form.watch([
+    "download.downloadThumb",
+    "download.downloadPoster",
+    "download.downloadFanart",
+    "download.downloadSceneImages",
+    "download.downloadTrailer",
+  ]) as [boolean | undefined, boolean | undefined, boolean | undefined, boolean | undefined, boolean | undefined];
+  const folderTemplate = String(form.watch("naming.folderTemplate") ?? "");
+  const successFileMove = Boolean(form.watch("behavior.successFileMove"));
+  const sharedDirectoryMode = isSharedDirectoryMode({ successFileMove, folderTemplate });
+
+  return (
+    <>
+      {sharedDirectoryMode && (
+        <div className="rounded-md border border-amber-500/30 bg-amber-500/10 px-3 py-2 text-xs text-amber-700">
+          当前为共享目录模式：多个影片会写入同一目录。保存时会校验 NFO 命名与剧照下载设置。
+        </div>
+      )}
+      <BoolField name="download.downloadThumb" label="下载横版缩略图" />
+      <BoolField name="download.downloadPoster" label="下载海报" />
+      {downloadPoster && (
+        <BoolField
+          name="download.tagBadges"
+          label="为封面添加标签角标"
+          description="按现有影片标签自动添加角标，当前支持中字、无码、破解、流出；仅处理本次新下载的海报。"
+        />
+      )}
+      <BoolField name="download.downloadFanart" label="下载背景图" />
+      <BoolField name="download.downloadSceneImages" label="下载剧照" />
+      <BoolField name="download.downloadTrailer" label="下载预告片" />
+      {downloadThumb && <BoolField name="download.keepThumb" label="保留已有横版缩略图" />}
+      {downloadPoster && <BoolField name="download.keepPoster" label="保留已有海报" />}
+      {downloadFanart && <BoolField name="download.keepFanart" label="保留已有背景图" />}
+      {downloadSceneImages && <BoolField name="download.keepSceneImages" label="保留已有剧照" />}
+      {downloadTrailer && <BoolField name="download.keepTrailer" label="保留已有预告片" />}
+    </>
+  );
+}
+
+export function NfoSection() {
+  const form = useFormContext<FieldValues>();
+  const generateNfo = Boolean(form.watch("download.generateNfo"));
+
+  return (
+    <>
+      <BoolField name="download.generateNfo" label="生成 NFO" />
+      {generateNfo && (
+        <>
+          <EnumField name="download.nfoNaming" label="NFO 文件命名" options={NFO_NAMING_OPTIONS} />
+          <BoolField name="download.keepNfo" label="保留已有 NFO" />
+        </>
+      )}
+    </>
+  );
+}
+
+function NamingPreview() {
+  const form = useFormContext<FieldValues>();
+  const previewValues = useWatch({
+    control: form.control,
+    name: NAMING_PREVIEW_FIELD_KEYS,
+  }) as unknown[];
+  const [previews, setPreviews] = useState<NamingPreviewItem[]>([]);
+  const [loading, setLoading] = useState(false);
+  const previewConfig = useMemo(() => {
+    const flatValues: Record<string, unknown> = {};
+    for (const [index, key] of NAMING_PREVIEW_FIELD_KEYS.entries()) {
+      flatValues[key] = previewValues[index];
+    }
+    return buildNamingPreviewConfig(flatValues);
+  }, [previewValues]);
+  const previewConfigRef = useRef(previewConfig);
+  const previewConfigKey = useMemo(() => JSON.stringify(previewConfig), [previewConfig]);
+
+  useEffect(() => {
+    previewConfigRef.current = previewConfig;
+  }, [previewConfig]);
+
+  useEffect(() => {
+    const requestKey = previewConfigKey;
+    let cancelled = false;
+    const timer = window.setTimeout(async () => {
+      setLoading(true);
+      try {
+        const result = await ipc.config.previewNaming(previewConfigRef.current as Partial<Configuration>);
+        if (!cancelled && requestKey === previewConfigKey) {
+          setPreviews(result.items);
+        }
+      } catch {
+        if (!cancelled && requestKey === previewConfigKey) {
+          setPreviews([]);
+        }
+      } finally {
+        if (!cancelled && requestKey === previewConfigKey) {
+          setLoading(false);
+        }
+      }
+    }, 120);
+
+    return () => {
+      cancelled = true;
+      window.clearTimeout(timer);
+    };
+  }, [previewConfigKey]);
+
+  return (
+    <div className="rounded-md border bg-muted/30 p-3">
+      <div className="mb-2 text-xs font-medium text-muted-foreground">命名预览</div>
+      <div className="space-y-2">
+        {previews.length === 0 && (
+          <div className="text-xs text-muted-foreground">{loading ? "生成预览中..." : "暂无预览"}</div>
+        )}
+        {previews.map((p) => (
+          <div key={p.label} className="text-xs">
+            <span className="mr-2 inline-block min-w-[4em] text-muted-foreground">{p.label}</span>
+            <span className="font-mono">
+              {p.folder}/{p.file}
+            </span>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
+
+export function NamingSection() {
+  const form = useFormContext<FieldValues>();
+  const folderTemplate = String(form.watch("naming.folderTemplate") ?? "");
+  const successFileMove = Boolean(form.watch("behavior.successFileMove"));
+  const sharedDirectoryMode = isSharedDirectoryMode({ successFileMove, folderTemplate });
+
+  return (
+    <>
+      <TextField name="naming.folderTemplate" label="文件夹模板" description={NAMING_TEMPLATE_DESCRIPTION} />
+      <TextField name="naming.fileTemplate" label="文件名模板" description={NAMING_TEMPLATE_DESCRIPTION} />
+      <EnumField
+        name="naming.assetNamingMode"
+        label="附属文件命名"
+        description="海报、横版缩略图、背景图与预告片的文件名规则。"
+        options={ASSET_NAMING_OPTIONS}
+      />
+      {sharedDirectoryMode && (
+        <div className="rounded-md border border-amber-500/30 bg-amber-500/10 px-3 py-2 text-xs text-amber-700">
+          当前文件夹模板不会为每部影片创建独立目录，属于共享目录模式。推荐默认使用 <code>{`{actor}/{number}`}</code>；
+          如需共享目录，保存时会校验相关命名规则。
+        </div>
+      )}
+      <TextField
+        name="naming.nfoTitleTemplate"
+        label="NFO 标题模板"
+        description="NFO 中 title 字段的格式。可用占位符：{number} {title} {originaltitle}"
+      />
+      <NamingPreview />
+      <NumberField name="naming.actorNameMax" label="演员名最大数量" min={1} max={20} />
+      <TextField name="naming.actorNameMore" label="演员名超出后缀" />
+      <BoolField
+        name="naming.actorFallbackToStudio"
+        label="演员为空时使用片商或卖家"
+        description="开启后，{actor} 在没有演员名时会回退到片商或卖家名称；如需显示来源，可在模板中使用 {actorFallbackPrefix}{actor}。"
+      />
+      <TextField name="naming.releaseRule" label="发行日期格式" />
+      <EnumField
+        name="naming.partStyle"
+        label="分盘样式"
+        description="分盘的视频在输出时保留原始后缀，或统一改写为 CD / PART / DISC 风格"
+        options={PART_STYLE_OPTIONS}
+      />
+      <NumberField name="naming.folderNameMax" label="文件夹名最大长度" min={10} max={255} />
+      <NumberField name="naming.fileNameMax" label="文件名最大长度" min={10} max={255} />
+      <TextField name="naming.cnwordStyle" label="中文字幕标记" />
+      <TextField name="naming.umrStyle" label="UMR 标记" />
+      <TextField name="naming.leakStyle" label="流出标记" />
+      <TextField name="naming.uncensoredStyle" label="无码标记" />
+      <TextField name="naming.censoredStyle" label="有码标记" />
+    </>
+  );
+}
+
+export function TranslateSection() {
+  const [testing, setTesting] = useState(false);
+  const form = useFormContext<FieldValues>();
+  const engine = useWatch({ control: form.control, name: "translate.engine" });
+  const isLLM = engine !== "google";
+
+  const handleTestLlm = async () => {
+    const input = {
+      llmModelName: String(form.getValues("translate.llmModelName") ?? ""),
+      llmApiKey: String(form.getValues("translate.llmApiKey") ?? ""),
+      llmBaseUrl: String(form.getValues("translate.llmBaseUrl") ?? ""),
+      llmTemperature: Number(form.getValues("translate.llmTemperature") ?? 0),
+    };
+
+    setTesting(true);
+    try {
+      const result = await ipc.translate.testLlm(input);
+      if (result.success) {
+        toast.success(result.message);
+      } else {
+        toast.error(result.message);
+      }
+    } catch (error) {
+      toast.error(`测试失败: ${error instanceof Error ? error.message : "未知错误"}`);
+    } finally {
+      setTesting(false);
+    }
+  };
+
+  return (
+    <>
+      <BaseField name="translate.enableTranslation" label="启用内容翻译">
+        {(field) => (
+          <div className="flex items-center gap-2">
+            {isLLM && (
+              <Button
+                type="button"
+                variant="outline"
+                size="sm"
+                className="h-7 text-xs"
+                onClick={handleTestLlm}
+                disabled={testing}
+              >
+                {testing ? (
+                  <>
+                    <Loader2 className="h-3 w-3 mr-1 animate-spin" /> 测试中...
+                  </>
+                ) : (
+                  "测试连通性"
+                )}
+              </Button>
+            )}
+            <FormControl>
+              <Switch checked={Boolean(field.value)} onCheckedChange={field.onChange} />
+            </FormControl>
+          </div>
+        )}
+      </BaseField>
+      <EnumField name="translate.engine" label="翻译引擎" options={TRANSLATE_ENGINE_OPTIONS} />
+      {isLLM && (
+        <>
+          <TextField name="translate.llmModelName" label="LLM 模型名称" />
+          <SecretField
+            name="translate.llmApiKey"
+            label="LLM 密钥（可选）"
+            description="默认 OpenAI 地址通常必须填写；本地或兼容服务是否需要密钥取决于服务端配置"
+          />
+          <UrlField
+            name="translate.llmBaseUrl"
+            label="LLM API 地址"
+            description={`默认值：${DEFAULT_LLM_BASE_URL}。本地常见示例：Ollama 用 http://127.0.0.1:11434/v1，LM Studio 用 http://127.0.0.1:1234/v1`}
+          />
+          <PromptFieldWrapper name="translate.llmPrompt" label="LLM 翻译提示词" />
+          <NumberField name="translate.llmTemperature" label="LLM 温度" min={0} max={2} step={0.1} />
+          <NumberField name="translate.llmMaxRetries" label="LLM 最大重试次数" min={1} max={20} />
+          <NumberField name="translate.llmMaxRequestsPerSecond" label="LLM 每秒最大请求数" min={1} max={100} />
+        </>
+      )}
+      <EnumField name="translate.targetLanguage" label="目标语言" options={LANGUAGE_OPTIONS} />
+    </>
+  );
+}
+
+export function PersonSyncSharedSection() {
+  return (
+    <div className="space-y-4 rounded-xl border bg-muted/10 p-4">
+      <div className="space-y-1">
+        <h4 className="text-sm font-medium">共享人物资料源</h4>
+        <p className="text-xs text-muted-foreground">
+          同时服务 Jellyfin 和 Emby。人物简介会按顺序选择一个质量达标的主资料源。
+        </p>
+      </div>
+      <ChipArrayFieldWrapper
+        name="personSync.personOverviewSources"
+        label="人物简介来源顺序"
+        options={ACTOR_OVERVIEW_SOURCE_OPTIONS}
+      />
+      <ChipArrayFieldWrapper
+        name="personSync.personImageSources"
+        label="人物头像来源顺序"
+        options={ACTOR_IMAGE_SOURCE_OPTIONS}
+      />
+    </div>
+  );
+}
+
+export function JellyfinSection() {
+  return (
+    <>
+      <UrlField name="jellyfin.url" label="Jellyfin 服务器地址" />
+      <CookieFieldWrapper name="jellyfin.apiKey" label="Jellyfin API Key" />
+      <TextField
+        name="jellyfin.userId"
+        label="Jellyfin 用户 ID"
+        description="必须是 UUID。用于人物列表读取，留空则按服务端默认处理。"
+      />
+      <BoolField
+        name="jellyfin.refreshPersonAfterSync"
+        label="同步后刷新人物"
+        description="同步简介或头像后，额外请求 Jellyfin 刷新人物元数据与图片。"
+      />
+      <BoolField
+        name="jellyfin.lockOverviewAfterSync"
+        label="同步后锁定人物简介"
+        description="写入简介后把 Overview 加入 LockedFields，降低被 Jellyfin 元数据刷新覆盖的概率。"
+      />
+    </>
+  );
+}
+
+export function EmbySection() {
+  return (
+    <>
+      <UrlField name="emby.url" label="Emby 服务器地址" />
+      <CookieFieldWrapper name="emby.apiKey" label="Emby API Key" />
+      <TextField name="emby.userId" label="Emby 用户 ID" description="用于人物列表读取，留空则按服务端默认处理。" />
+      <BoolField
+        name="emby.refreshPersonAfterSync"
+        label="同步后刷新人物"
+        description="同步简介或头像后，额外请求 Emby 刷新人物元数据与图片。"
+      />
+    </>
+  );
+}
+
+export function ShortcutsSection() {
+  return (
+    <>
+      <ShortcutField name="shortcuts.startOrStopScrape" label="开始/停止刮削" description="示例: S" />
+      <ShortcutField name="shortcuts.retryScrape" label="重新刮削" description="示例: R" />
+      <ShortcutField name="shortcuts.deleteFile" label="删除文件" description="示例: D" />
+      <ShortcutField name="shortcuts.deleteFileAndFolder" label="删除文件及文件夹" description="示例: ⇧ + D" />
+      <ShortcutField name="shortcuts.openFolder" label="打开所在目录" description="示例: F" />
+      <ShortcutField name="shortcuts.editNfo" label="编辑 NFO" description="示例: E" />
+      <ShortcutField name="shortcuts.playVideo" label="播放视频" description="示例: P" />
+    </>
+  );
+}
+
+interface UiSectionProps {
+  initialUseCustomTitleBar: boolean;
+}
+
+export function UiSection({ initialUseCustomTitleBar }: UiSectionProps) {
+  const [relaunching, setRelaunching] = useState(false);
+  const form = useFormContext<FieldValues>();
+  const currentUseCustomTitleBar = Boolean(useWatch({ control: form.control, name: "ui.useCustomTitleBar" }) ?? true);
+  const titleBarChanged = currentUseCustomTitleBar !== initialUseCustomTitleBar;
+  const canRelaunch = titleBarChanged && !form.formState.isDirty;
+
+  const handleRelaunch = async () => {
+    if (form.formState.isDirty) {
+      toast.info("请先保存设置，再重启应用");
+      return;
+    }
+
+    setRelaunching(true);
+    try {
+      await ipc.app.relaunch();
+    } catch (error) {
+      setRelaunching(false);
+      toast.error(`重启失败: ${error instanceof Error ? error.message : "未知错误"}`);
+    }
+  };
+
+  return (
+    <>
+      <BoolField name="ui.showLogsPanel" label="显示日志面板" />
+      <BaseField name="ui.useCustomTitleBar" label="使用自定义标题栏" description="切换后需要重启应用">
+        {(field) => (
+          <div className="flex items-center gap-2">
+            <Button
+              type="button"
+              variant={titleBarChanged ? "default" : "outline"}
+              size="sm"
+              className="h-7 rounded-lg text-xs"
+              disabled={!canRelaunch || relaunching}
+              onClick={handleRelaunch}
+            >
+              {relaunching ? <Loader2 className="h-3 w-3 animate-spin" /> : <RotateCcw className="h-3 w-3" />}
+              重启应用
+            </Button>
+            <FormControl>
+              <Switch checked={Boolean(field.value ?? true)} onCheckedChange={field.onChange} />
+            </FormControl>
+          </div>
+        )}
+      </BaseField>
+      <BoolField name="ui.hideDock" label="隐藏 Dock 图标" />
+      <BoolField name="ui.hideMenu" label="隐藏菜单栏" />
+      <BoolField name="ui.hideWindowButtons" label="隐藏窗口按钮" />
+    </>
+  );
+}
+
+export function BehaviorSection() {
+  return (
+    <>
+      <BoolField name="behavior.successFileMove" label="成功后移动文件" />
+      <BoolField name="behavior.failedFileMove" label="失败后移动文件" />
+      <BoolField name="behavior.successFileRename" label="成功后重命名文件" />
+      <BoolField name="behavior.deleteEmptyFolder" label="删除空文件夹" />
+      <BoolField name="behavior.scrapeSoftlinkPath" label="刮削软链接目录" />
+      <BoolField name="behavior.saveLog" label="保存日志到文件" />
+    </>
+  );
+}
